requirements.txt
django-headless-cms==1.1.0
  Django==4.2.17
    asgiref==3.8.1
    sqlparse==0.5.3
  django-admin-interface==0.29.2
    django-colorfield==0.11.0
      pillow==11.0.0
    python-slugify==8.0.4
      text-unidecode==1.3
  django-admin-sortable2==2.2.4
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
  django-filter==24.3
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
  django-import-export==4.3.3
    diff-match-patch==20241021
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
    tablib==3.7.0
  django-localized-fields==6.7
    deprecation==2.0.7
      packaging==24.2
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
    django-postgres-extra==2.0.8
      Django==4.2.17
        asgiref==3.8.1
        sqlparse==0.5.3
      python-dateutil==2.9.0.post0
        six==1.17.0
  django-reversion==5.1.0
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
  django-solo==2.4.0
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
  djangorestframework==3.15.2
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
  drf-spectacular==0.28.0
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
    djangorestframework==3.15.2
      Django==4.2.17
        asgiref==3.8.1
        sqlparse==0.5.3
    inflection==0.5.1
    jsonschema==4.23.0
      attrs==24.3.0
      jsonschema-specifications==2024.10.1
        referencing==0.35.1
          attrs==24.3.0
          rpds-py==0.22.3
      referencing==0.35.1
        attrs==24.3.0
        rpds-py==0.22.3
      rpds-py==0.22.3
    PyYAML==6.0.2
    uritemplate==4.1.1
  martor==1.6.45
    bleach==6.2.0
      webencodings==0.5.1
    Django==4.2.17
      asgiref==3.8.1
      sqlparse==0.5.3
    Markdown==3.5.2
    requests==2.32.3
      certifi==2024.12.14
      charset-normalizer==3.4.1
      idna==3.10
      urllib3==2.3.0
    tzdata==2024.2
  psycopg==3.2.3
    typing_extensions==4.12.2
  Unidecode==1.3.8
pipdeptree==2.24.0
  packaging==24.2
  pip==24.3.1
python-dotenv==1.0.1
setuptools==66.1.1
